+++ example/a.ts
@@ -1,20 +1,11 @@
 import { Elysia } from '../src'
 import { req } from '../test/utils'
-import { newWebsocket, wsClosed, wsOpen } from '../test/ws/utils'
 
 const app = new Elysia()
+	.get('/ws', () => 'hi')
 	.ws('/ws', {
 		message() {}
 	})
-	.get('/ws', () => 'hi')
 	.listen(3000)
 
-console.log(app.fetch.toString())
-
-// const ws = newWebsocket(app.server!)
-
-// await wsOpen(ws)
-// await wsClosed(ws)
-
-const response = await app.handle(req('/ws')).then((x) => x.text())
-console.log(response) // .toBe('hi')
+console.log(app.routes[0].compile().toString())